requirements.txt
Django==2.0.2
django-filter==1.0.4
djangorestframework==3.7.7
gunicorn==19.7.1
psycopg2==2.7.1
django-storages==1.6.3
boto
kong_oauth==0.1

requests
responses
faker

ipdb
ipython

coverage
codacy-coverage
